правки по код ревью
