+++ src/main/java/school/faang/bjs286069/Book.java
@@ -1,0 +1,17 @@
+package school.faang.bjs286069;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@EqualsAndHashCode
+public class Book {
+
+    private String title;
+    private String author;
+    private int year;
+}
